requirements.txt
numpy
scipy
matplotlib
pydmd
scikit-learn
pyyaml